requirements/requirements.txt
transformers 
torch 
google-cloud-bigquery 
python-dotenv